+++ src/components/quote.js
@@ -1,0 +1,17 @@
+import React from 'react';
+
+function Quote() {
+  return (
+    <div id="quotepage">
+      <div id="quote">
+        <p>Calculators can only calculate - they cannot do mathematics.</p>
+      </div>
+      <div id="author">
+        <p>John A. Van de Walle</p>
+      </div>
+
+    </div>
+  );
+}
+
+export default Quote;
